+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased - 2020-10-05
 
 - Added AIP Random which will assign a random Absurd or AIP personality
+- Rush techs earlier
 
 ## v1.5.3 - 2020-07-12